:recycle: refactor(api): move legistar calls behind tRPC route
Direct client-side Legistar API calls bypass server caching, rate
limiting, and fail with CORS on Expo web. Added legistar tRPC router
and updated LocalBillsSection to use it.
