+++ packages/api/src/root.ts
@@ -2,6 +2,7 @@ import { authRouter } from "./router/auth";
 import { civicRouter } from "./router/civic";
 import { contentRouter } from "./router/content";
 import { electionsRouter } from "./router/elections";
+import { legistarRouter } from "./router/legistar";
 import { postRouter } from "./router/post";
 import { videoRouter } from "./router/video";
 import { createTRPCRouter } from "./trpc";
@@ -9,6 +10,7 @@ import { createTRPCRouter } from "./trpc";
 export const appRouter = createTRPCRouter({
   auth: authRouter,
   civic: civicRouter,
+  legistar: legistarRouter,
   post: postRouter,
   content: contentRouter,
   video: videoRouter,
